pyproject.toml
[project]
name = "playwright-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "drissionpage>=4.1.0.17",
    "fake-useragent>=2.0.3",
    "playwright>=1.50.0",
    "python-dotenv>=1.0.1",
]
